+++ changelog.md
@@ -1,5 +1,7 @@
 # Changes to zhixuhao's original implementation
 
+* `geneTrainNpY` was refactored and renamed `TrainAugmented`
+
 * Added `thresholdImages` to threshold masks(mostly). Please see [pyautocv](https://github.com/Nelson-Gon/pyautocv)
 for a more general and flexible way to manipulate images.